+++ src/lib/helpers/uint-packing.ts
