+++ src/lib/terrain/voxelmap/voxelsRenderable/voxelsRenderableFactory/merged/vertex-data1-encoder.ts
@@ -1,6 +1,6 @@
 import type * as THREE from '../../../../../three-usage';
 import { type VoxelsChunkSize } from '../../../i-voxelmap';
-import { PackedUintFactory, type PackedUintFragment } from '../uint-packing';
+import { PackedUintFactory, type PackedUintFragment } from '../../../../../helpers/uint-packing';
 
 class VertexData1Encoder {
     private readonly packedUintFactory = new PackedUintFactory(32);
