+++ src/lib/terrain/voxelmap/voxelsRenderable/voxelsRenderableFactory/merged/vertex-data2-encoder.ts
@@ -1,4 +1,4 @@
-import { PackedUintFactory } from '../uint-packing';
+import { PackedUintFactory } from '../../../../../helpers/uint-packing';
 
 class VertexData2Encoder {
     private readonly packedUintFactory = new PackedUintFactory(32);
